--- a/Util.h
+++ b/Util.h
@@ -11,6 +11,7 @@
 #include <iterator>
 #include <ctime>
 #include <vector>
+#include <omp.h>
 
 struct PhasingElement{
     // i.e. 0|1  or   1|0
@@ -18,6 +19,33 @@ struct PhasingElement{
     int block;
 };
 typedef std::map<std::string,PhasingElement> PhasingResult;
+typedef std::map<std::string,PhasingResult> ChrPhasingResult;
+
+/**
+ * Merges phasing results from all chromosomes into one result.
+ *
+ * @param allChrPhasingResults Contains phasing results for each chromosome.
+ * @param mergedPhasingResult A map to store the combined phasing results.
+ *
+ * The function combines the phasing elements from each chromosome into mergedPhasingResult.
+ */
+void mergeAllChrPhasingResult(const ChrPhasingResult& allChrPhasingResults, PhasingResult& mergedPhasingResult);
+
+/**
+ * Configures thread counts for chromosome processing and BAM parsing.
+ *
+ * @param defaultThreads Initial thread count for chromosome processing.
+ * @param availableThreads  Total threads available for processing
+ * @param chrnumThreads Set to the number of threads for chromosome processing.
+ * @param bamParsernumThreads Set to the number of threads for BAM parsing.
+ *
+ * Depending on availableThreads, this function adjusts chrnumThreads and bamParsernumThreads 
+ * for optimized parallel processing. The thread allocation strategy changes based on 
+ * whether availableThreads is less than or equal to 24, between 25 and 48, or greater than 48.
+ * This function's thread allocation strategy is designed considering the typical count 
+ * of human chromosomes.
+ */
+void setNumThreads(const int& defaultChrThreads,const int& availableThreads,  int& chrnumThreads, int& bamParsernumThreads);
 
 // use for parsing
 struct Variant{